requirements.txt
google-generativeai>=0.5.0
requests>=2.31.0
pytest>=7.4.0
beautifulsoup4>=4.12.0
playwright>=1.47.0
google-genai>=0.3.0
model-context-protocol>=0.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3
pypdf>=4.2.0
